movie-api-postgresSQL/config.py
from decouple import config
#Decouple it help me to get env variables

class Config:
    SECRET_KEY=config('SECRET_KEY')

class DevelopmentConfig(Config):
    DEBUG = True

config = {
    'development': DevelopmentConfig
}