movie-api-postgresSQL/requirements.txt
flask~=3.0.3
flask-cors
psycopg2~=2.9.9
python-decouple
python-dotenv